+++ src/pem-parser.c
@@ -10,18 +10,18 @@ uint32_t base64_value[128];
 
 void base64_value_init(void){
 
-    for(ssize_t i = 0; i < 128; i++){
+    for(uint32_t i = 0; i < 128; i++){
         base64_value[i] = 64;
     }
 
     uint32_t counter = 0;
-    for(ssize_t i = 'A'; i <= 'Z'; i++){
+    for(uint32_t i = 'A'; i <= 'Z'; i++){
         base64_value[i] = counter; counter++;
     }
-    for(ssize_t i = 'a'; i <= 'z'; i++){
+    for(uint32_t i = 'a'; i <= 'z'; i++){
         base64_value[i] = counter; counter++;
     }
-    for(ssize_t i = '0'; i <= '9'; i++){
+    for(uint32_t i = '0'; i <= '9'; i++){
         base64_value[i] = counter; counter++;
     }
 
@@ -31,27 +31,27 @@ void base64_value_init(void){
 }
 
 
-ssize_t base64_decode(char *encoded, uint8_t *decoded){
-    ssize_t n = strlen(encoded);
+uint32_t base64_decode(char *encoded, uint8_t *decoded){
+    uint32_t n = strlen(encoded);
 
     if(n % 4){
-        fprintf(stderr, "[ERROR] base64 encoded text length must be a multiple of four, length: %lu.\n",  n);
+        fprintf(stderr, "[ERROR] base64 encoded text length must be a multiple of four, length: %u.\n",  n);
         return 0;
     }
 
-    for(ssize_t i = 0; i < n; i++){
+    for(uint32_t i = 0; i < n; i++){
         if(encoded[i] < 0 || encoded[i] >= 128 || base64_value[(size_t) encoded[i]] == 64){
             fprintf(stderr, "[ERROR] Invalid base64 character (%c).\n", encoded[i]);
             return 0;
         }
     }
 
-    ssize_t decoded_size = 0;
-    for(ssize_t i = 0; i < n; i += 4){
+    uint32_t decoded_size = 0;
+    for(uint32_t i = 0; i < n; i += 4){
 
         uint32_t value = 0;
-        ssize_t padding_count = 0;
-        for(ssize_t j  = 0; j < 4; j++){
+        uint32_t padding_count = 0;
+        for(uint32_t j  = 0; j < 4; j++){
             value <<= 6;
             value += base64_value[(size_t) encoded[i + j]];
             if(encoded[i + j] == '=') padding_count++;
@@ -59,7 +59,7 @@ ssize_t base64_decode(char *encoded, uint8_t *decoded){
  
 
         decoded_size += 3 - padding_count;
-        for(ssize_t j = 0; j < 3; j++){
+        for(uint32_t j = 0; j < 3; j++){
             if(j >= padding_count){
                 decoded_size--;
                 decoded[decoded_size] = value % 256;
@@ -86,7 +86,7 @@ uint8_t parse_pem_footer(char *header){
 uint8_t parse_pem(FILE *file, ParseTree *parseTree){
 
     uint8_t der[MAX_DER_FILE_SIZE];
-    ssize_t der_size = 0;
+    uint32_t der_size = 0;
 
     char buffer[MAX_FILE_BUFFER_SIZE];
     uint8_t decoded[MAX_FILE_BUFFER_SIZE];
@@ -103,13 +103,13 @@ uint8_t parse_pem(FILE *file, ParseTree *parseTree){
         }
 
         buffer[strlen(buffer) - 1] = '\0';
-        ssize_t decoded_size = base64_decode(buffer, decoded);
+        uint32_t decoded_size = base64_decode(buffer, decoded);
         if(decoded_size == 0){
             printf("[ERROR] Invalid PEM content.\n");
             return 1;
         }
 
-        for(ssize_t i = 0; i < decoded_size; i++){
+        for(uint32_t i = 0; i < decoded_size; i++){
             der[der_size] = decoded[i];
             der_size++;
         }
